+++ sender/src/main/java/com/philips/Utils.java
@@ -1,4 +1,4 @@
-package com.main;
+package com.philips;
 
 public class Utils {
